requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
click==7.1.2
dj-database-url==0.5.0
Django==2.2.17
djangorestframework==3.12.2
gunicorn==20.0.4
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pytz==2020.1
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
